+++ Dracula/Actions/ABayDroneTrigger.cs
@@ -1,25 +1,37 @@
 ﻿using Nickel;
 using System.Collections.Generic;
+using System.Linq;
 
 namespace Shockah.Dracula;
 
-public sealed class APositionalDroneTrigger : CardAction
+public sealed class ABayDroneTrigger : CardAction
 {
-	public required int WorldX;
+	public required bool FromPlayer;
 
 	public override void Begin(G g, State s, Combat c)
 	{
 		base.Begin(g, s, c);
 		timer = 0;
-		if (!c.stuff.TryGetValue(WorldX, out var @object))
-			return;
-		c.QueueImmediate(@object.GetActions(s, c) ?? []);
+
+		var ship = FromPlayer ? s.ship : c.otherShip;
+		c.QueueImmediate(
+			Enumerable.Range(0, ship.parts.Count)
+				.Where(i => ship.parts[i].type == PType.missiles && ship.parts[i].active)
+				.Select(i => new APositionalDroneTrigger { WorldX = ship.x + i })
+		);
 	}
 
 	public override List<Tooltip> GetTooltips(State s)
 	{
-		if (s.route is Combat combat && combat.stuff.TryGetValue(WorldX, out var @object))
-			@object.hilight = 2;
+		foreach (var i in Enumerable.Range(0, s.ship.parts.Count))
+		{
+			if (s.ship.parts[i].type != PType.missiles || !s.ship.parts[i].active)
+				continue;
+			s.ship.parts[i].hilight = true;
+
+			if (s.route is Combat combat && combat.stuff.TryGetValue(s.ship.x + i, out var @object))
+				@object.hilight = 2;
+		}
 
 		return [
 			new GlossaryTooltip($"{ModEntry.Instance.Package.Manifest.UniqueName}::Action::DroneTrigger")