+++ Dracula/Actions/ATooltipAction.cs
@@ -1,11 +1,9 @@
-﻿using Newtonsoft.Json;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 
 namespace Shockah.Dracula;
 
 internal sealed class ATooltipAction : ADummyAction
 {
-	[JsonProperty]
 	public List<Tooltip>? Tooltips;
 
 	public override void Begin(G g, State s, Combat c)